programmatic navigation router
create router using useRouter() helper method from vue

watch() when value > 1200, navigate to other page

difference between router.push() vs router.replace()
- https://stackoverflow.com/a/65933466 below
I like to use .replace() when i have a definitive action and i dont want the user to be able to come back to this page.

Example: i have a product detail page with a delete button to ultimately deletes the product. When i click the button i do router.replace back to the list of all products so when the user uses the back button in the browser he is not able to go back to the details page so he cannot see or even click again the delete button as this product does not exist anymore.